+++ src/annbatch/samplers/__init__.py
@@ -1,4 +1,5 @@
-from ._chunk_sampler import ChunkSampler, DistributedRandomSampler
+from ._chunk_sampler import ChunkSampler
+from ._distributed_random_sampler import DistributedRandomSampler
 from ._random_sampler import RandomSampler
 from ._sequential_sampler import SequentialSampler
